requirements.txt
numpy
opencv-python
torch
facenet-pytorch
scikit-learn
flask
flask-cors
pillow
matplotlib